Refactor distributed Client
New client implementation with message chunking
