+++ crates/lunatic-distributed-api/src/lib.rs
@@ -3,7 +3,10 @@ use std::{future::Future, sync::Arc, time::Duration};
 use anyhow::{anyhow, Result};
 use lunatic_common_api::{get_memory, IntoTrap};
 use lunatic_distributed::{
-    distributed::message::{ClientError, Spawn, Val},
+    distributed::{
+        client::{EnvironmentId, NodeId, ProcessId, SendParams, SpawnParams},
+        message::{Spawn, Val},
+    },
     DistributedCtx,
 };
 use lunatic_error_api::ErrorCtx;
@@ -276,39 +279,29 @@ where
 
         log::debug!("Spawn on node {node_id}, mod {module_id}, fn {function}, params {params:?}");
 
-        let (process_or_error_id, ret) = match state
-            .distributed()?
-            .node_client
-            .spawn(
-                node_id,
-                Spawn {
-                    environment_id: state.environment_id(),
-                    function: function.to_string(),
-                    module_id,
-                    params,
-                    config,
-                },
-            )
-            .await
-        {
-            Ok(process_id) => (process_id, 0),
-            Err(error) => {
-                let (code, message): (u32, String) = match error {
-                    ClientError::Unexpected(cause) => Err(anyhow!(cause)),
-                    ClientError::NodeNotFound => Ok((1, "Node does not exist.".to_string())),
-                    ClientError::ModuleNotFound => Ok((2, "Module does not exist.".to_string())),
-                    ClientError::Connection(cause) => Ok((9027, cause)),
-                    _ => Err(anyhow!("unreachable")),
-                }?;
-                (
-                    caller
+        let spawn_params = SpawnParams {
+            env: EnvironmentId(state.environment_id()),
+            src: ProcessId(state.id()),
+            node: NodeId(node_id),
+            spawn: Spawn {
+                environment_id: state.environment_id(),
+                function: function.to_string(),
+                module_id,
+                params,
+                config,
+            },
+        };
+        let (process_or_error_id, ret) =
+            match state.distributed()?.node_client.spawn(spawn_params).await {
+                Ok(process_id) => (process_id, 0),
+                Err(message) => {
+                    let error_id = caller
                         .data_mut()
                         .error_resources_mut()
-                        .add(anyhow!(message)),
-                    code,
-                )
-            }
-        };
+                        .add(anyhow!(message));
+                    (error_id, 1)
+                }
+            };
 
         memory
             .write(
@@ -364,20 +357,17 @@ where
             }
 
             let state = caller.data();
-            match state
-                .distributed()?
-                .node_client
-                .message_process(node_id, state.environment_id(), process_id, tag, buffer)
-                .await
-            {
+            let send_params = SendParams {
+                env: EnvironmentId(state.environment_id()),
+                src: ProcessId(state.id()),
+                node: NodeId(node_id),
+                dest: ProcessId(process_id),
+                tag,
+                data: buffer,
+            };
+            match state.distributed()?.node_client.send(send_params).await {
                 Ok(_) => Ok(0),
-                Err(error) => match error {
-                    ClientError::Unexpected(cause) => Err(anyhow!(cause)),
-                    ClientError::ProcessNotFound => Ok(1),
-                    ClientError::NodeNotFound => Ok(2),
-                    ClientError::Connection(_) => Ok(9027),
-                    _ => Err(anyhow!("unreachable")),
-                },
+                Err(cause) => Err(anyhow!(cause)),
             }
         } else {
             Err(anyhow!("Only Message::Data can be sent across nodes."))
@@ -437,19 +427,17 @@ where
             }
 
             let state = caller.data();
-            let code = match state
-                .distributed()?
-                .node_client
-                .message_process(node_id, state.environment_id(), process_id, tag, buffer)
-                .await
-            {
+            let send_params = SendParams {
+                env: EnvironmentId(state.environment_id()),
+                src: ProcessId(state.id()),
+                node: NodeId(node_id),
+                dest: ProcessId(process_id),
+                tag,
+                data: buffer,
+            };
+            let code = match state.distributed()?.node_client.send(send_params).await {
                 Ok(_) => Ok(0),
-                Err(error) => match error {
-                    ClientError::ProcessNotFound => Ok(1),
-                    ClientError::NodeNotFound => Ok(2),
-                    ClientError::Unexpected(cause) => Err(anyhow!(cause)),
-                    _ => Err(anyhow!("unreachable")),
-                },
+                Err(error) => Err(anyhow!(error)),
             }?;
 
             if code != 0 {
