+++ crates/lunatic-distributed/src/congestion/mod.rs
@@ -1,62 +1,18 @@
 use std::{
     collections::VecDeque,
-    sync::{
-        atomic::{self, AtomicU64, AtomicUsize},
-        Arc,
-    },
+    sync::{atomic, Arc},
 };
 
-use anyhow::{anyhow, Result};
-use dashmap::DashMap;
+use anyhow::Result;
 use tokio::sync::{
     mpsc::{self, error::TryRecvError, Receiver, Sender},
     RwLock,
 };
 
 use crate::{
-    control,
-    distributed::message::{Request, Spawn},
+    distributed::{self, client::ProcessId},
     quic, NodeInfo,
 };
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub struct EnvironmentId(pub u64);
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub struct ProcessId(pub u64);
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub struct NodeId(pub u64);
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub struct MessageId(pub u64);
-
-pub struct SendParams {
-    pub env: EnvironmentId,
-    pub src: ProcessId,
-    pub node: NodeId,
-    pub dest: ProcessId,
-    pub tag: Option<i64>,
-    pub data: Vec<u8>,
-}
-
-pub struct SpawnParams {
-    pub env: EnvironmentId,
-    pub src: ProcessId,
-    pub node: NodeId,
-    pub spawn: Spawn,
-}
-
-pub struct MessageCtx {
-    message_id: MessageId,
-    env: EnvironmentId,
-    src: ProcessId,
-    node: NodeId,
-    dest: ProcessId,
-    chunk_id: AtomicU64,
-    offset: AtomicUsize,
-    data: Vec<u8>,
-}
 
 pub struct MessageChunk {
     src: ProcessId,
@@ -67,139 +23,10 @@ pub struct MessageChunk {
     data: bytes::Bytes,
 }
 
-// Receiving part of the message queue
-type BufRx = RwLock<Receiver<MessageCtx>>;
-
-// TODO: replace distributed::Client
-#[derive(Clone)]
-pub struct Client {
-    inner: Arc<Inner>,
-}
-
-pub struct Inner {
-    control_client: control::Client,
-    node_client: quic::Client,
-    next_message_id: AtomicU64,
-    // Across Environments and ProcessId's track message queues
-    buf_rx: DashMap<EnvironmentId, DashMap<ProcessId, BufRx>>,
-    // Sending part of the message queue
-    buf_tx: DashMap<(EnvironmentId, ProcessId), Sender<MessageCtx>>,
-    // Holds the message while its being chunked
-    in_progress: DashMap<(EnvironmentId, ProcessId), MessageCtx>,
-    nodes_queues: DashMap<NodeId, Sender<MessageChunk>>,
-}
-
-impl Client {
-    fn next_message_id(&self) -> MessageId {
-        MessageId(
-            self.inner
-                .next_message_id
-                .fetch_add(1, atomic::Ordering::Relaxed),
-        )
-    }
-
-    async fn new_message(
-        &self,
-        env: EnvironmentId,
-        src: ProcessId,
-        node: NodeId,
-        dest: ProcessId,
-        data: Vec<u8>,
-    ) -> Result<()> {
-        // Lazy initialize process message buffers
-        let tx = match self.inner.buf_tx.get(&(env, src)) {
-            Some(tx) => tx,
-            None => {
-                let (send, recv) = tokio::sync::mpsc::channel(100); // TODO: configuration
-                match self.inner.buf_rx.get(&env) {
-                    Some(env_queue) => {
-                        env_queue.insert(src, RwLock::new(recv));
-                    }
-                    None => {
-                        let queue = DashMap::new();
-                        queue.insert(src, RwLock::new(recv));
-                        self.inner.buf_rx.insert(env, queue);
-                    }
-                };
-                self.inner.buf_tx.insert((env, src), send);
-                self.inner.buf_tx.get(&(env, src)).unwrap()
-            }
-        };
-
-        let node_manager_exists = self.inner.nodes_queues.get(&node).is_none();
-
-        if node_manager_exists {
-            let node_info = self
-                .inner
-                .control_client
-                .node_info(node.0)
-                .ok_or_else(|| anyhow!(""))?;
-            let (send, recv) = tokio::sync::mpsc::channel(100);
-            tokio::spawn(node_connection_manager(NodeConnectionManager {
-                streams: 10,
-                node_info,
-                client: self.inner.node_client.clone(),
-                message_chunks: recv,
-            }));
-            self.inner.nodes_queues.insert(node, send);
-        }
-
-        match tx
-            .send(MessageCtx {
-                message_id: self.next_message_id(),
-                env,
-                src,
-                node,
-                dest,
-                offset: AtomicUsize::new(0),
-                chunk_id: AtomicU64::new(0),
-                data,
-            })
-            .await
-        {
-            Ok(_) => (),
-            Err(_) => log::error!("lunatic::distributed::client::send"),
-        };
-        Ok(())
-    }
-
-    // TODO: how to detect process is dead?
-    pub fn remove_process_resources(&self, env: EnvironmentId, process_id: ProcessId) {
-        self.inner.buf_tx.remove(&(env, process_id));
-    }
-
-    // Send distributed message
-    pub async fn send(&self, params: SendParams) -> Result<()> {
-        let message = Request::Message {
-            environment_id: params.env.0,
-            process_id: params.dest.0,
-            tag: params.tag,
-            data: params.data,
-        };
-        let data = match rmp_serde::to_vec(&message) {
-            Ok(data) => data,
-            Err(_) => unreachable!("lunatic::distributed::client::send serialize_message"),
-        };
-        self.new_message(params.env, params.src, params.node, params.dest, data)
-            .await
-    }
-
-    // Send distributed spawn message
-    pub async fn spawn(&self, params: SpawnParams) -> Result<()> {
-        let message = Request::Spawn(params.spawn);
-        let data = match rmp_serde::to_vec(&message) {
-            Ok(data) => data,
-            Err(_) => unreachable!("lunatic::distributed::client::spawn serialize_message"),
-        };
-        self.new_message(params.env, params.src, params.node, ProcessId(0), data)
-            .await
-    }
-}
-
 // TODO: move to configuration
 const CHUNK_SIZE: usize = 1024;
 
-pub async fn congestion_control_worker(state: Client) -> ! {
+pub async fn congestion_control_worker(state: distributed::Client) -> ! {
     loop {
         for env in state.inner.buf_rx.iter() {
             let mut disconected = vec![];
